% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcalheatmap.R
\name{rcalheatmapOutput}
\alias{rcalheatmapOutput}
\title{Widget output function for use in Shiny}
\usage{
rcalheatmapOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

