% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcalheatmap.R
\name{renderRcalheatmap}
\alias{renderRcalheatmap}
\title{Widget render function for use in Shiny}
\usage{
renderRcalheatmap(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

